refactor: TokenPair 클래스 분리 및 관련 참조 수정
- AuthTokenService 내 TokenPair 클래스 분리를 통해 재사용성 개선
- TokenPair를 user.domain 패키지로 이동하여 도메인 객체로 재정의
- OAuthCodeRepository 및 RedisOAuthCodeRepository에서 TokenPair 참조 수정
- 관련 테스트 코드 내 TokenPair 패키지 경로 수정
- 코드 간소화 및 유지보수성 강화 목적
